feat(pool-transactions): cleanup and minor improvements
